refactor: 리뷰 enum 들을 tag 패키지로 이동
