+++ src/main/java/com/workat/api/review/dto/ReviewDto.java
@@ -1,6 +1,6 @@
 package com.workat.api.review.dto;
 
-import com.workat.domain.review.BaseReviewType;
+import com.workat.domain.tag.BaseTag;
 
 import io.swagger.annotations.ApiModelProperty;
 import lombok.AccessLevel;
@@ -22,7 +22,7 @@ public class ReviewDto {
 		this.count = count;
 	}
 
-	public static ReviewDto of(BaseReviewType reviewType, long count) {
+	public static ReviewDto of(BaseTag reviewType, long count) {
 		return new ReviewDto(ReviewTypeDto.of(reviewType), count);
 	}
 }
