--- src/main/java/com/workat/domain/review/BaseReviewType.java
@@ -1,8 +1,0 @@
-package com.workat.domain.review;
-
-public interface BaseReviewType {
-
-	String getName();
-
-	String getContent();
-}
