+++ src/main/java/com/workat/domain/tag/BaseTag.java
@@ -1,0 +1,8 @@
+package com.workat.domain.tag;
+
+public interface BaseTag {
+
+	String getName();
+
+	String getContent();
+}
