+++ src/test/java/com/workat/api/review/service/ReviewServiceTest.java
@@ -30,11 +30,11 @@ import com.workat.domain.map.entity.Location;
 import com.workat.domain.map.entity.LocationCategory;
 import com.workat.domain.map.http.dto.KakaoLocalDataDto;
 import com.workat.domain.map.repository.location.LocationRepository;
-import com.workat.domain.review.CafeReviewType;
 import com.workat.domain.review.entity.CafeReview;
 import com.workat.domain.review.entity.RestaurantReview;
 import com.workat.domain.review.repository.CafeReviewRepository;
 import com.workat.domain.review.repository.RestaurantReviewRepository;
+import com.workat.domain.tag.CafeReviewType;
 import com.workat.domain.user.entity.UserProfile;
 import com.workat.domain.user.entity.Users;
 import com.workat.domain.user.job.DepartmentType;
